Renamed file containing Selection type and useSelection() method
